Refactor consai_robot_controller (#191)
* Update debug message

* Add detection_extractor

* Add constraint_parser

* Update CosntraintParser

* Add tactic_control_ball

* Remove unused functions from field_info_parser

* Add tactic_obstacle_avoidance

* Remove unused functions from field_info_parser

* Add obstacle observer

* Refactoring

* Fix for lint

* update cmake list

* Move some files to obstacle dir

* Make tactic and tools dir

* refactoring

* refactor kick

* Change field_info_parser test to obstacle observer test

* fix kick speed
