+++ consai_robot_controller/CMakeLists.txt
@@ -29,12 +29,17 @@ include_directories(include)
 
 # Controller Component
 add_library(controller_component SHARED
-  src/ball_boy_tactics.cpp
   src/controller_component.cpp
   src/field_info_parser.cpp
-  src/geometry_tools.cpp
-  src/control_tools.cpp
   src/visualization_data_handler.cpp
+  src/detection_extractor.cpp
+  src/constraint_parser.cpp
+  src/obstacle/obstacle_observer.cpp
+  src/tactic/ball_boy_tactics.cpp
+  src/tactic/tactic_control_ball.cpp
+  src/tactic/tactic_obstacle_avoidance.cpp
+  src/tools/control_tools.cpp
+  src/tools/geometry_tools.cpp
 )
 target_compile_definitions(controller_component
   PRIVATE "CONSAI_ROBOT_CONTROLLER_BUILDING_DLL")
@@ -103,17 +108,16 @@ if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
   ament_lint_auto_find_test_dependencies()
   find_package(ament_cmake_gtest)
-  ament_add_gtest(control_tools_test test/control_tools_test.cpp src/control_tools.cpp)
-  ament_add_gtest(field_info_parser_test
-    test/field_info_parser_test.cpp
-    src/ball_boy_tactics.cpp
-    src/geometry_tools.cpp
-    src/field_info_parser.cpp)
-  ament_target_dependencies(field_info_parser_test
-    consai_msgs
-    rclcpp
-    robocup_ssl_msgs
+  ament_add_gtest(control_tools_test
+    test/control_tools_test.cpp
+    src/tools/control_tools.cpp
   )
+
+  ament_add_gtest(obstacle_observer_test
+    test/obstacle_observer_test.cpp
+  )
+  target_link_libraries(obstacle_observer_test controller_component)
+
   ament_add_gtest(obstacle_environment_test test/obstacle_environment_test.cpp)
 endif()
