+++ consai_robot_controller/include/consai_robot_controller/controller_component.hpp
@@ -30,6 +30,8 @@
 #include "consai_robot_controller/field_info_parser.hpp"
 #include "consai_robot_controller/visibility_control.h"
 #include "consai_robot_controller/visualization_data_handler.hpp"
+#include "consai_robot_controller/detection_extractor.hpp"
+#include "consai_robot_controller/obstacle/obstacle_observer.hpp"
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp_action/rclcpp_action.hpp"
 #include "robocup_ssl_msgs/msg/geometry_data.hpp"
@@ -93,7 +95,10 @@ private:
   std::vector<std::shared_ptr<GoalHandleRobotControl>> goal_handle_;
   std::vector<State> last_world_vel_;
 
-  consai_robot_controller::FieldInfoParser parser_;
+  std::shared_ptr<consai_robot_controller::FieldInfoParser> parser_;
+  std::shared_ptr<parser::DetectionExtractor> detection_extractor_;
+  std::shared_ptr<obstacle::ObstacleObserver> obstacle_observer_;
+
   rclcpp::Subscription<TrackedFrame>::SharedPtr sub_detection_tracked_;
   rclcpp::Subscription<GeometryData>::SharedPtr sub_geometry_;
   rclcpp::Subscription<NamedTargets>::SharedPtr sub_named_targets_;
