+++ consai_robot_controller/include/consai_robot_controller/obstacle/obstacle_typedef.hpp
@@ -1,4 +1,4 @@
-// Copyright 2021 Roots
+// Copyright 2023 Roots
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,18 +12,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef CONSAI_ROBOT_CONTROLLER__CONTROL_TOOLS_HPP_
-#define CONSAI_ROBOT_CONTROLLER__CONTROL_TOOLS_HPP_
+#ifndef CONSAI_ROBOT_CONTROLLER__OBSTACLE__OBSTACLE_TYPEDEF_HPP_
+#define CONSAI_ROBOT_CONTROLLER__OBSTACLE__OBSTACLE_TYPEDEF_HPP_
 
-#include <complex>
-#include <algorithm>
-
-namespace control_tools
+namespace obstacle
 {
 
-double velocity_contol_tanh(const double diff, const double range, const double max_vel);
-double angular_velocity_contol_sin(const double diff, const double max_vel);
+class Point
+{
+public:
+  Point(const double x, const double y)
+  : x_(x), y_(y)
+  {
+  }
 
-}  // namespace control_tools
+  double x() const
+  {
+    return x_;
+  }
 
-#endif  // CONSAI_ROBOT_CONTROLLER__CONTROL_TOOLS_HPP_
+  double y() const
+  {
+    return y_;
+  }
+
+private:
+  double x_;
+  double y_;
+};
+
+using Radius = double;
+using Position = Point;
+
+}  // namespace obstacle
+
+#endif  // CONSAI_ROBOT_CONTROLLER__OBSTACLE__OBSTACLE_TYPEDEF_HPP_
