+++ consai_robot_controller/include/consai_robot_controller/tactic/ball_boy_tactics.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef CONSAI_ROBOT_CONTROLLER__BALL_BOY_TACTICS_HPP_
-#define CONSAI_ROBOT_CONTROLLER__BALL_BOY_TACTICS_HPP_
+#ifndef CONSAI_ROBOT_CONTROLLER__TACTIC__BALL_BOY_TACTICS_HPP_
+#define CONSAI_ROBOT_CONTROLLER__TACTIC__BALL_BOY_TACTICS_HPP_
 
 #include <chrono>
 #include <map>
@@ -95,4 +95,4 @@ private:
 
 }  // namespace tactics
 
-#endif  // CONSAI_ROBOT_CONTROLLER__BALL_BOY_TACTICS_HPP_
+#endif  // CONSAI_ROBOT_CONTROLLER__TACTIC__BALL_BOY_TACTICS_HPP_
