+++ consai_robot_controller/src/tactic/ball_boy_tactics.cpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "consai_robot_controller/ball_boy_tactics.hpp"
-#include "consai_robot_controller/geometry_tools.hpp"
+#include "consai_robot_controller/tactic/ball_boy_tactics.hpp"
+#include "consai_robot_controller/tools/geometry_tools.hpp"
 
 namespace tactics
 {
