+++ consai_robot_controller/src/tools/geometry_tools.cpp
@@ -16,7 +16,7 @@
 #include <cmath>
 #include <iostream>
 
-#include "consai_robot_controller/geometry_tools.hpp"
+#include "consai_robot_controller/tools/geometry_tools.hpp"
 
 namespace geometry_tools
 {
